pyproject.toml
[project]
name = "sirius"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic",
    "httpx",
    "sentence-transformers",
    "transformers>=4.40",
    "scikit-learn>=1.3",
    "numpy",
    "hydra-core>=1.3",
]

[project.optional-dependencies]
dev = ["pytest"]
gpu = ["torch", "torchvision", "torchaudio", "accelerate"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that call external APIs (skipped by default)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sirius"]

# CPU-only PyTorch by default; install with `uv sync --extra gpu` for CUDA builds.
[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "extra != 'gpu'" },
    { index = "pytorch-cuda", marker = "extra == 'gpu'" },
]
torchvision = [
    { index = "pytorch-cpu", marker = "extra != 'gpu'" },
    { index = "pytorch-cuda", marker = "extra == 'gpu'" },
]
torchaudio = [
    { index = "pytorch-cpu", marker = "extra != 'gpu'" },
    { index = "pytorch-cuda", marker = "extra == 'gpu'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cuda"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
